python-service/pyproject.toml
[project]
name = "ib-quiz-python-service"
version = "0.1.0"
description = "OCR and Math solver microservice for IB Quiz App"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pix2tex>=0.1.2",
    "sympy>=1.13.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.18",
    "pillow>=11.0.0",
    "numpy>=2.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "httpx>=0.28.0",
]
